backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
SpeechRecognition==3.10.4
gTTS==2.5.3
python-dotenv==1.0.1
requests==2.32.3
psutil==6.0.0
PyAudio==0.2.14
pywinauto==0.6.8
pyautogui==0.9.54
pyperclip==1.9.0 